requirements.txt
adblock @ file:///build/python-adblock/src/python-adblock-0.6.0/target/wheels/adblock-0.0.0-cp37-abi3-manylinux_2_34_x86_64.whl
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
apsw==3.43.1.0
arandr==0.1.11
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
Brlapi==0.8.5
Brotli==1.1.0
brotlicffi==1.1.0.0
btrfsutil==6.7.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.0
coverage==7.4.1
cryptography==42.0.5
css-parser==1.0.10
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.6
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnspython==2.5.0
docutils==0.20.1
envycontrol==3.3.1
evdev==1.7.0
executing==2.0.0
fastjsonschema==2.19.1
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.13.1
flann==1.9.2
fonttools==4.50.0
fqdn==1.5.1
h11==0.14.0
html2text==2020.1.16
html5-parser==0.4.11
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
ifaddr==0.2.0
importlib_metadata==5.1.0
inflate64==1.0.0
inflect==7.0.0
injector==0.21.0
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.3.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyterlab==4.1.5
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keyring==24.3.0
kiwisolver==1.4.5
lensfun==0.3.4
libclang==16.0.6
libtorrent==2.0.9
lit==17.0.6.dev0
lockfile==0.12.2
louis==3.29.0
lxml==5.1.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.8
meson==1.4.0
mistune==3.0.2
moddb==0.10.0
more-itertools==10.2.0
mpi4py==3.1.5
msgpack==1.0.5
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest_asyncio==1.5.9
netifaces==0.11.0
netsnmp-python==1.0a1
nh3==0.2.14
notebook==7.0.6
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.535.133
nvitop==1.3.2
ordered-set==4.1.0
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
paramiko==2.11.1
parso==0.8.3
pdftotext==2.2.2
peewee==3.17.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==4.2.0
ply==3.11
pooch==1.8.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
py3nvml==0.2.7
py7zr==0.21.0
pybcj==1.0.2
pybind11==2.11.1
pycairo==1.26.0
pychm==0.8.6
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.12.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyGObject==3.48.1
PyNaCl==1.4.0
pynvml==11.5.0
PyOpenGL==3.1.6
pyparsing==3.1.2
pyppmd==1.1.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0
pyqtgraph==0.13.3
pyrate-limiter==3.6.0
pyrsistent==0.19.3
pyserial==3.5
PySide2==5.15.13
PySide6==6.6.2
python-dateutil==2.9.0
python-json-logger==2.0.7
python-xlib==0.33
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.2
pyzstd==0.15.10
qtconsole==5.5.1
QtPy==2.4.1
qutebrowser==3.1.0
ranger-fm==1.9.3
reactivex==4.0.4
readme-renderer==42.0
referencing==0.33.0
Reflector==2023.6.28.0.36.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
scour==0.38.2
SecretStorage==3.3.3
Send2Trash==1.8.2
setproctitle==1.3.3
sgmllib3k==1.0.0
shiboken2==5.15.13
shiboken6==6.6.2
shiboken6-generator==6.6.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
TBB==0.2
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.2.0
thrift==0.19.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.14.2
trash-cli==0.23.11.10
trove-classifiers==2024.3.25
twine==4.0.2
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
uc-micro-py==1.0.3
unrardll==0.1.7
uri-template==1.3.0
urllib3==1.26.18
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.25.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wxPython==4.2.1
xxhash==3.4.1
yt-dlp==2024.3.10
zeroconf==0.63.0
zipp==3.17.0
zstandard==0.22.0
